Turrets and Particles added.

+++ Assets/Scripts/Datas/PlayerMovementData/MovementData.cs
@@ -7,7 +7,7 @@ namespace Datas.PlayerMovementData
     
     public class MovementData : ScriptableObject
     {
-        public int Speed;
+        public float Speed;
 
     }
 }
